+++ site/config.js
@@ -1,5 +1,6 @@
 window.DOCS_FORGE_CONFIG = Object.freeze({
   policyVersion: "2026-07-29",
+  consentMode: "external",
   formEndpoint: "",
   consentEndpoint: "",
   posthogKey: "",